ALL DONE
please notice that I did not think parallel, sorry for that. It was for time reason... I will though consider it for the HALFSPACE MASS EX. Also, i did not find any efficient idea to deal with multicollinear variables. One could eventually calculate a correlation matrix & use a warning if a matrix element (other than the diagonal ones) is bigger than a certain threshold...